Extract 8 playlist helpers into modules/output_playlists.py + drop 1 dead one
Second refactor pass at modules/output.py.  Moved eight helpers that
deal with two closely-related concerns:

  playlist file entry normalization:
    * _normalize_playlist_file_entry_for_output
    * _parse_playlist_file_entries_value
    * _format_playlist_file_entries

  playlist library selection:
    * _ordered_selected_libraries
    * _library_names_in_output_order
    * _playlist_libraries_from_library_toggles
    * _legacy_playlist_libraries_from_settings
    * _legacy_playlist_libraries_for_selected_libraries

YAGNI bonus: dropped _format_playlist_files -- a thin one-line wrapper
around _format_playlist_file_entries with zero call sites anywhere in
the codebase.

Dependencies: the new module needs only modules.persistence and
modules.output_values (which we split out in the previous PR).

Public surface: output.py explicitly re-imports the eight names with
'# noqa: F401' so historical call sites like output._format_playlist_file_entries
continue to work if any tests reach for them.

Impact:
- modules/output.py: 3670 -> 3538 (-132 / -3.6%)
- modules/output_playlists.py: 172 lines new
- Dead code purged: 3 lines (_format_playlist_files)
- Zero behavior change

All 1077 tests pass. Ruff + black clean.
